requirements.txt
langchain
langchain-community
langchain-openai
tavily-python
openai
faiss-cpu
python-dotenv
tiktoken
fastembed

